cambios visuales, add politica de acceso a repos

+++ app/Http/Controllers/RepositoryController.php
@@ -3,24 +3,20 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\RepositoryRequest;
-use Illuminate\Http\Request;
 use App\Models\Repository;
 
 class RepositoryController extends Controller
 {
-    public function index(Request $request)
+    public function index()
     {
-        $repositories = $request->user()->repositories;
+        $repositories = auth()->user()->repositories;
 
         return view('repositories.index', compact('repositories'));
     }
 
-    public function show(Request $request, Repository $repository)
+    public function show(Repository $repository)
     {
-        if($request->user()->id != $repository->user_id)
-        {
-            abort(403);
-        }
+        $this->authorize('pass', $repository);
         return view('repositories.show', compact('repository'));
     }
 
@@ -36,21 +32,15 @@ class RepositoryController extends Controller
         return redirect()->route('repositories.index');
     }
 
-    public function edit(Request $request, Repository $repository)
+    public function edit(Repository $repository)
     {
-        if($request->user()->id != $repository->user_id)
-        {
-            abort(403);
-        }
+        $this->authorize('pass', $repository);
         return view('repositories.edit', compact('repository'));
     }
 
     public function update(RepositoryRequest $request, Repository $repository)
     {
-        if($request->user()->id != $repository->user_id)
-        {
-            abort(403);
-        }
+        $this->authorize('pass', $repository);
 
         $repository->update($request->all());
 
@@ -59,6 +49,8 @@ class RepositoryController extends Controller
 
     public function destroy(Repository $repository)
     {
+        $this->authorize('pass', $repository);
+
         $repository->delete();
 
         return redirect()->route('repositories.index');
